+++ packages/monsterdog/HARMONIZE_COLLECTIVE/index.js
@@ -1,0 +1,3 @@
+module.exports = function(params){
+  return { ok: true, action: 'HARMONIZE_COLLECTIVE', params };
+};