+++ packages/monsterdog/OPTIMIZE_PERFORMANCE/index.js
@@ -1,0 +1,3 @@
+module.exports = function(params){
+  return { ok: true, action: 'OPTIMIZE_PERFORMANCE', params };
+};